requirements.txt
pystray
Pillow
requests
pyyaml
